# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiftEase API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Recipient {
  id: number;
  name: string;
  /** @nullable */
  relationship: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRecipientBody {
  name: string;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateRecipientBody {
  name?: string;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Occasion {
  id: number;
  /** @nullable */
  recipientId: number | null;
  title: string;
  /** @nullable */
  date: string | null;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOccasionBody {
  /** @nullable */
  recipientId?: number | null;
  title: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateOccasionBody {
  /** @nullable */
  recipientId?: number | null;
  title?: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Gift {
  id: number;
  /** @nullable */
  recipientId: number | null;
  /** @nullable */
  occasionId: number | null;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  url: string | null;
  purchased: boolean;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGiftBody {
  /** @nullable */
  recipientId?: number | null;
  /** @nullable */
  occasionId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateGiftBody {
  /** @nullable */
  recipientId?: number | null;
  /** @nullable */
  occasionId?: number | null;
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ToggleGiftPurchasedBody {
  purchased: boolean;
}

export type ListGiftsParams = {
  recipientId?: number;
  occasionId?: number;
};