# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiftEase API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all recipients
 */
export const ListRecipientsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  relationship: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListRecipientsResponse = zod.array(ListRecipientsResponseItem);

/**
 * @summary Create a recipient
 */
export const CreateRecipientBody = zod.object({
  name: zod.string(),
  relationship: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a recipient
 */
export const GetRecipientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRecipientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  relationship: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a recipient
 */
export const UpdateRecipientParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateRecipientBody = zod.object({
  name: zod.string().optional(),
  relationship: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateRecipientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  relationship: zod.string().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a recipient
 */
export const DeleteRecipientParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all occasions
 */
export const ListOccasionsResponseItem = zod.object({
  id: zod.number(),
  recipientId: zod.number().nullable(),
  title: zod.string(),
  date: zod.coerce.date().nullable(),
  budget: zod.number().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListOccasionsResponse = zod.array(ListOccasionsResponseItem);

/**
 * @summary Create an occasion
 */
export const CreateOccasionBody = zod.object({
  recipientId: zod.number().nullish(),
  title: zod.string(),
  date: zod.coerce.date().nullish(),
  budget: zod.number().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Update an occasion
 */
export const UpdateOccasionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOccasionBody = zod.object({
  recipientId: zod.number().nullish(),
  title: zod.string().optional(),
  date: zod.coerce.date().nullish(),
  budget: zod.number().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateOccasionResponse = zod.object({
  id: zod.number(),
  recipientId: zod.number().nullable(),
  title: zod.string(),
  date: zod.coerce.date().nullable(),
  budget: zod.number().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an occasion
 */
export const DeleteOccasionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all gifts
 */
export const ListGiftsQueryParams = zod.object({
  recipientId: zod.coerce.number().optional(),
  occasionId: zod.coerce.number().optional(),
});

export const ListGiftsResponseItem = zod.object({
  id: zod.number(),
  recipientId: zod.number().nullable(),
  occasionId: zod.number().nullable(),
  name: zod.string(),
  description: zod.string().nullable(),
  price: zod.number().nullable(),
  url: zod.string().nullable(),
  purchased: zod.boolean(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListGiftsResponse = zod.array(ListGiftsResponseItem);

/**
 * @summary Create a gift idea
 */
export const CreateGiftBody = zod.object({
  recipientId: zod.number().nullish(),
  occasionId: zod.number().nullish(),
  name: zod.string(),
  description: zod.string().nullish(),
  price: zod.number().nullish(),
  url: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Update a gift
 */
export const UpdateGiftParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateGiftBody = zod.object({
  recipientId: zod.number().nullish(),
  occasionId: zod.number().nullish(),
  name: zod.string().optional(),
  description: zod.string().nullish(),
  price: zod.number().nullish(),
  url: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateGiftResponse = zod.object({
  id: zod.number(),
  recipientId: zod.number().nullable(),
  occasionId: zod.number().nullable(),
  name: zod.string(),
  description: zod.string().nullable(),
  price: zod.number().nullable(),
  url: zod.string().nullable(),
  purchased: zod.boolean(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a gift
 */
export const DeleteGiftParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle gift purchased status
 */
export const ToggleGiftPurchasedParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleGiftPurchasedBody = zod.object({
  purchased: zod.boolean(),
});

export const ToggleGiftPurchasedResponse = zod.object({
  id: zod.number(),
  recipientId: zod.number().nullable(),
  occasionId: zod.number().nullable(),
  name: zod.string(),
  description: zod.string().nullable(),
  price: zod.number().nullable(),
  url: zod.string().nullable(),
  purchased: zod.boolean(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});