# lib/api-zod/src/generated/types/createOccasionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiftEase API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOccasionBody {
  /** @nullable */
  recipientId?: number | null;
  title: string;
  /** @nullable */
  date?: Date | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  notes?: string | null;
}