# lib/api-zod/src/generated/types/gift.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiftEase API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Gift {
  id: number;
  /** @nullable */
  recipientId: number | null;
  /** @nullable */
  occasionId: number | null;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  url: string | null;
  purchased: boolean;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  updatedAt: Date;
}