# lib/api-zod/src/generated/types/updateRecipientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiftEase API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRecipientBody {
  name?: string;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  notes?: string | null;
}